refactors all designs into js modules and importing shared functions from a new util.js file
