+++ designs/circles.js
@@ -1,5 +1,7 @@
-function drawCircles(ctx, seed) {
-  let rng = sfc32(bits(seed),bits(seed),bits(seed),bits(seed));
+import { sfc32, bits } from "../utils.js";
+
+export function draw(ctx, seed) {
+  let rng = sfc32(bits(seed), bits(seed), bits(seed), bits(seed));
   let points = [];
   let disc = (x, y, r) => {
     ctx.beginPath();
@@ -15,10 +17,10 @@ function drawCircles(ctx, seed) {
   for (let i = 0; i < 10; ++i) {
     let phi = rng() * Math.PI * 2;
     let r = Math.sqrt(rng()) * 0.92;
-    points.push({x: Math.cos(phi) * r, y: Math.sin(phi) * r, phi});
+    points.push({ x: Math.cos(phi) * r, y: Math.sin(phi) * r, phi });
   }
-  ctx.strokeStyle = 'gray';
-  points.forEach(({x, y, phi}) => {
+  ctx.strokeStyle = "gray";
+  points.forEach(({ x, y, phi }) => {
     let cx = Math.cos(phi) * 0.96;
     let cy = Math.sin(phi) * 0.96;
     line(x / 2 + 0.5, y / 2 + 0.5, cx / 2 + 0.5, cy / 2 + 0.5);
@@ -26,9 +28,8 @@ function drawCircles(ctx, seed) {
     ctx.arc(0.5, 0.5, 0.48, phi - Math.PI / 72, phi + Math.PI / 72);
     ctx.stroke();
   });
-  ctx.fillStyle = 'black';
-  points.forEach(({x, y}) => {
+  ctx.fillStyle = "black";
+  points.forEach(({ x, y }) => {
     disc(x / 2 + 0.5, y / 2 + 0.5, 0.03);
   });
 }
-addSchema('Circles', drawCircles);
